FIRM28: Refactor init files, add HAL_MspInit(), create MCU struct
