+++ Src/common/stm32/mcu/init_stm32g474.c
@@ -11,11 +11,6 @@
 
 #ifdef STM32G474xx
 
-void clock_init(void) {
-	SystemClock_Config();
-	GPIOClock_Config();
-}
-
 /**
   * @brief System Clock Configuration
   * @retval None
@@ -70,21 +65,6 @@ void SystemClock_Config(void)
   {
     Error_Handler();
   }
-}
-
-/**
-  * @brief GPIO Clock Configuration
-  * @retval None
-  */
-void GPIOClock_Config(void) {
-	// Need to use #ifdef here for just the H7 because
-	// __HAL_RCC_GPIOH_CLK_ENABLE() is not defined in the G4 HAL
-
-  /* GPIO Ports Clock Enable */
-  __HAL_RCC_GPIOC_CLK_ENABLE();
-  __HAL_RCC_GPIOF_CLK_ENABLE();
-  __HAL_RCC_GPIOA_CLK_ENABLE();
-  __HAL_RCC_GPIOB_CLK_ENABLE();
 }
 
 #endif
